requirements.txt
python-telegram-bot>=20.7
yt-dlp>=2024.4.9
pydantic-settings>=2.0.0
